Move SqlClient implementations to tests
With the exception fo the AdapterBackedSqlClient, All of our existing
implementations are used exclusively for running tests until they
are eventually replaced by an adapter integration.

We move these to the tests package to allow us to remove the various
SQLAlchemy and other SQL engine dependencies from Metricflow's core
dependency list.
