+++ metricflow/test/fixtures/sql_clients/duckdb.py
@@ -15,9 +15,9 @@ from metricflow.protocols.sql_client import SqlEngine
 from metricflow.sql.render.duckdb_renderer import DuckDbSqlQueryPlanRenderer
 from metricflow.sql.render.sql_plan_renderer import SqlQueryPlanRenderer
 from metricflow.sql.sql_bind_parameters import SqlBindParameters
-from metricflow.sql_clients.common_client import SqlDialect
-from metricflow.sql_clients.sqlalchemy_dialect import SqlAlchemySqlClient
 from metricflow.sql_request.sql_request_attributes import SqlJsonTag, SqlRequestTagSet
+from metricflow.test.fixtures.sql_clients.common_client import SqlDialect
+from metricflow.test.fixtures.sql_clients.sqlalchemy_dialect import SqlAlchemySqlClient
 
 logger = logging.getLogger(__name__)
