+++ metricflow/test/fixtures/sql_clients/redshift.py
@@ -9,8 +9,8 @@ from typing_extensions import override
 from metricflow.protocols.sql_client import SqlEngine
 from metricflow.sql.render.redshift import RedshiftSqlQueryPlanRenderer
 from metricflow.sql.render.sql_plan_renderer import SqlQueryPlanRenderer
-from metricflow.sql_clients.common_client import SqlDialect, not_empty
-from metricflow.sql_clients.sqlalchemy_dialect import SqlAlchemySqlClient
+from metricflow.test.fixtures.sql_clients.common_client import SqlDialect, not_empty
+from metricflow.test.fixtures.sql_clients.sqlalchemy_dialect import SqlAlchemySqlClient
 
 logger = logging.getLogger(__name__)
